requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
boto3==1.28.3
botocore==1.31.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.3.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.4
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
promise==2.3
psycopg==3.1.9
psycopg-binary==3.1.9
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
pytest-spec==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
